moved action to module
