+++ openpype/modules/python_console_interpreter/__init__.py
@@ -1,0 +1,8 @@
+from .module import (
+    PythonInterpreterAction
+)
+
+
+__all__ = (
+    "PythonInterpreterAction",
+)
